Fixed formatting and naming issues
