+++ src/site/shared/containers/Header/Left/SideBarToggleButton.tsx
@@ -8,7 +8,7 @@ export const SideBarToggleButton = () => {
     const dispatch = useSharedDispatch();
 
     return (
-            <div>
+        <div>
             <span title="Side Bar" role="button" tabIndex={0}
                   onClick={() => dispatch(ToggleSideNav())}>&#9776;</span>
         </div>
