+++ src/site/shared/containers/Header/Left/index.scss
@@ -81,22 +81,20 @@
     }
     &__history {
         margin-top: math.div($header-height - $header-history-height, 2) - $header-border-width;
-        margin-right: 10px;
+        margin-right: 7px;
 
         cursor: pointer;
 
         background-color: transparent;
 
-        & > button {
-            & > img {
-                height: $header-history-height;
-                filter: invert(.55);
-            }
+        img {
+            height: $header-history-height;
+            filter: invert(.55);
         }
     }
     &__lock {
         margin-top: math.div($header-height - $header-lock-height, 2) - $header-border-width;
-        margin-right: 5px;
+        margin-right: 7px;
 
         cursor: pointer;
