+++ src/site/shared/containers/Header/_constants.scss
@@ -1,3 +1,3 @@
 $header-button-height: 30px;
-$header-history-height: 30px;
-$header-lock-height: 24px;
+$header-history-height: 24px;
+$header-lock-height: 18px;
